+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed incorrect `Unknown macro` error when arguments are empty.
+- Fixed formatting long lines when column length + 1 === wordWrapColumn due to `;`
 
 ## [0.7.1] - 2025-12-02